requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.2.0
faiss-cpu==1.9.0.post1
numpy==1.24.3
scikit-learn==1.3.2
streamlit==1.28.1
python-multipart==0.0.6
pydantic==2.5.0
tqdm==4.66.1